fix[MINI-5635]: add custom permission test cases
